peel off Reporter into its own file
